+++ deebot_client/commands/xml/__init__.py
@@ -10,8 +10,8 @@ from .battery import GetBatteryInfo
 from .charge import Charge
 from .charge_state import GetChargeState
 from .clean_logs import GetCleanLogs
+from .clean_speed import GetCleanSpeed, SetCleanSpeed
 from .error import GetError
-from .fan_speed import GetCleanSpeed, SetCleanSpeed
 from .life_span import GetLifeSpan
 from .play_sound import PlaySound
 from .pos import GetChargerPos, GetPos
@@ -40,16 +40,16 @@ __all__ = [
 _COMMANDS: list[type[XmlCommand]] = [
     GetBatteryInfo,
 
-    Charge,
-
     GetChargeState,
+
+    Charge,
 
     GetCleanLogs,
 
-    GetError,
-
     GetCleanSpeed,
     SetCleanSpeed,
+
+    GetError,
 
     GetLifeSpan,
