+++ tests/commands/xml/test_clean_speed.py
